requirements.txt
Pillow>=10.0.0
pillow-heif>=0.15.0
PyYAML>=6.0 